Image uploading

+++ src/routes/editor/protected/home.svelte
@@ -1,17 +1,22 @@
 <script lang="ts" context="module">
 	import { goto } from '$app/navigation';
 	import { session } from '$app/stores';
+	import Uploader from '$lib/components/editor/uploader.svelte';
 	import Metadata from '$lib/components/metadata.svelte';
-	import { getAllCharacters, listPagePaths, pagePathToDocId } from '$lib/content';
+	import { getAllCharacters } from '$lib/content/characters';
+	import { listPagePaths, pagePathToDocId } from '$lib/content/pages';
+	import { listContentObjects } from '$lib/editor';
 	import type { CharacterObject } from '$lib/types/character';
 	import type { Load } from '@sveltejs/kit';
-	export const load: Load = async () => {
+	export const load: Load = async ({ session }) => {
 		const chars = await getAllCharacters(true);
 		const pagePaths = await listPagePaths();
+		const imageObjects = await listContentObjects(session.editor.key, 'images/');
 		return {
 			props: {
 				chars,
-				pagePaths
+				pagePaths,
+				imageObjects
 			}
 		};
 	};
@@ -20,6 +25,7 @@
 <script lang="ts">
 	export let chars: Map<string, CharacterObject>;
 	export let pagePaths: string[];
+	export let imageObjects: { name: string }[];
 
 	function deleteItem(collection: string, document: string): void {
 		if (
@@ -28,26 +34,47 @@
 			) != document
 		)
 			return;
-		window
-			.fetch(`/api/edit`, {
-				method: 'DELETE',
-				body: JSON.stringify({
-					editorKey: $session.editor.key,
-					collection,
-					document
-				})
+		fetch(`/api/edit`, {
+			method: 'DELETE',
+			body: JSON.stringify({
+				editorKey: $session.editor.key,
+				collection,
+				document
 			})
-			.then((resp) => {
-				if (resp.status == 200) {
-					window.location.reload();
-				} else {
-					console.error('Delete failed');
-				}
-			});
+		}).then((resp) => {
+			if (resp.status == 200) {
+				window.location.reload();
+			} else {
+				console.error('Delete failed');
+			}
+		});
+	}
+
+	function deleteObject(fileName: string): void {
+		if (
+			window.prompt(
+				`CONFIRM DELETION: "${fileName}"?\nType '${fileName}' to continue.\n\nNote: File may take a day to be removed completely.`
+			) != fileName
+		)
+			return;
+		fetch(`/api/files`, {
+			method: 'DELETE',
+			body: JSON.stringify({
+				editorKey: $session.editor.key,
+				name: fileName
+			})
+		}).then((resp) => {
+			if (resp.status == 200) {
+				window.location.reload();
+			} else {
+				console.error('Delete failed');
+			}
+		});
 	}
 
 	let selChar: string;
 	let selPage: string;
+	let selImg: string;
 </script>
 
 <Metadata title="Editor Home" description="Landing page for editors" />
@@ -83,6 +110,24 @@
 		</div>
 		<div>
 			<button on:click={() => goto(`./pages/!new`)}>Add Page</button>
+		</div>
+	</div>
+	<div class="images">
+		<h2>Images</h2>
+		<div>
+			<select bind:value={selImg}>
+				{#each imageObjects as imgObj}
+					<option value={imgObj.name}>{imgObj.name}</option>
+				{/each}
+			</select>
+			<button on:click={() => deleteObject(selImg)}>Delete</button>
+		</div>
+		<div style="padding-top: 10px;">
+			<Uploader
+				allowedTypes="image/png,image/jpeg"
+				filePrefix="images"
+				success={() => window.location.reload()}
+			/>
 		</div>
 	</div>
 </div>
